refactor: move sri plugin to uni-builder
