+++ .changeset/friendly-experts-smoke.md
@@ -1,0 +1,9 @@
+---
+'@rsbuild/uni-builder': patch
+'@rsbuild/webpack': patch
+'@rsbuild/doctor-core': patch
+'@rsbuild/document': patch
+'@rsbuild/shared': patch
+---
+
+refactor: move sri plugin to uni-builder
