+++ packages/compat/uni-builder/src/types.ts
@@ -94,8 +94,22 @@ export type UniBuilderExtraConfig = {
   };
 };
 
+export type SriOptions = {
+  hashFuncNames?: [string, ...string[]];
+  enabled?: 'auto' | true | false;
+  hashLoading?: 'eager' | 'lazy';
+};
+
 export type UniBuilderWebpackConfig = RsbuildWebpackConfig &
-  UniBuilderExtraConfig;
+  UniBuilderExtraConfig & {
+    security?: {
+      /**
+       * Adding an integrity attribute (`integrity`) to sub-resources introduced by HTML allows the browser to
+       * verify the integrity of the introduced resource, thus preventing tampering with the downloaded resource.
+       */
+      sri?: SriOptions | boolean;
+    };
+  };
 
 export type UniBuilderRspackConfig = RsbuildRspackConfig &
   UniBuilderExtraConfig;
