+++ packages/compat/webpack/package.json
@@ -65,8 +65,7 @@
     "terser-webpack-plugin": "5.3.9",
     "ts-loader": "9.4.4",
     "tsconfig-paths-webpack-plugin": "4.1.0",
-    "webpack": "^5.89.0",
-    "webpack-subresource-integrity": "5.1.0"
+    "webpack": "^5.89.0"
   },
   "devDependencies": {
     "@rsbuild/test-helper": "workspace:*",
