--- packages/document/docs/zh/shared/config/security/checkSyntax.md
@@ -1,13 +1,0 @@
-- **类型：**
-
-```ts
-type CheckSyntax =
-  | boolean
-  | {
-      targets?: string[];
-      exclude?: RegExp | RegExp[];
-      ecmaVersion?: EcmaVersion;
-    };
-```
-
-- **默认值：** `false`
