+++ packages/shared/src/types/config/security.ts
@@ -1,9 +1,3 @@
-export type SriOptions = {
-  hashFuncNames?: [string, ...string[]];
-  enabled?: 'auto' | true | false;
-  hashLoading?: 'eager' | 'lazy';
-};
-
 export interface SecurityConfig {
   /**
    * Adding an nonce attribute to sub-resources introduced by HTML allows the browser to
